requirements.txt
alembic==1.9.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
